+++ configs/eslint.js
@@ -1,6 +1,7 @@
 /* eslint-disable sort-keys */
 
 module.exports = {
+  extends: ['plugin:rut/recommended'],
   overrides: [
     {
       files: 'styles.ts',