+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
+- [@aesthetic/core-react](./packages/core-react/CHANGELOG.md)
 - [@aesthetic/react](./packages/react/CHANGELOG.md)
-- [@aesthetic/react-ui](./packages/react-ui/CHANGELOG.md)